Address comments on response format
Also add debug dump of htp requests
