+++ go/dbproxy/local_table.go
@@ -15,9 +15,7 @@ package dbproxy
 
 import (
 	"almond-cloud/sql"
-	"errors"
 	"net/http"
-	"strconv"
 
 	"github.com/gin-gonic/gin"
 	"gorm.io/gorm"
@@ -40,7 +38,7 @@ func localTableGetAll(c *gin.Context) {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 		return
 	}
-	c.JSON(http.StatusOK, gin.H{"data": rows})
+	c.JSON(http.StatusOK, gin.H{"result": "ok", "data": rows})
 }
 
 func localTableGetOne(c *gin.Context) {
@@ -64,29 +62,7 @@ func localTableGetOne(c *gin.Context) {
 		}
 		return
 	}
-	c.JSON(http.StatusOK, gin.H{"data": row})
-}
-func parseUserID(c *gin.Context) (int64, error) {
-	userID, err := strconv.ParseInt(c.Param("userid"), 10, 64)
-	if err != nil {
-		return 0, err
-	}
-	if userID == 0 {
-		return 0, errors.New("userId must be non-zero")
-	}
-	return userID, nil
-}
-
-func parseKey(c *gin.Context) (*sql.Key, error) {
-	userID, err := parseUserID(c)
-	if err != nil {
-		return nil, err
-	}
-	uniqueID := c.Param("uniqueid")
-	if len(uniqueID) == 0 {
-		return nil, errors.New("uniqueid must be set")
-	}
-	return &sql.Key{UniqueID: uniqueID, UserID: userID}, nil
+	c.JSON(http.StatusOK, gin.H{"result": "ok", "data": row})
 }
 
 func localTableDeleteOne(c *gin.Context) {
@@ -110,7 +86,7 @@ func localTableDeleteOne(c *gin.Context) {
 		}
 		return
 	}
-	c.JSON(http.StatusOK, gin.H{"data": true})
+	c.JSON(http.StatusOK, gin.H{"result": "ok", "data": true})
 }
 
 func localTableInsertOne(c *gin.Context) {
@@ -137,5 +113,5 @@ func localTableInsertOne(c *gin.Context) {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 		return
 	}
-	c.JSON(http.StatusOK, gin.H{"data": true})
+	c.JSON(http.StatusOK, gin.H{"result": "ok", "data": true})
 }
