+++ go/dbproxy/sync_table.go
@@ -39,7 +39,7 @@ func syncTableGetAll(c *gin.Context) {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 		return
 	}
-	c.JSON(http.StatusOK, gin.H{"data": rows})
+	c.JSON(http.StatusOK, gin.H{"result": "ok", "data": rows})
 }
 
 func syncTableGetOne(c *gin.Context) {
@@ -63,7 +63,7 @@ func syncTableGetOne(c *gin.Context) {
 		}
 		return
 	}
-	c.JSON(http.StatusOK, gin.H{"data": row})
+	c.JSON(http.StatusOK, gin.H{"result": "ok", "data": row})
 }
 
 func syncTableGetRaw(c *gin.Context) {
@@ -83,7 +83,7 @@ func syncTableGetRaw(c *gin.Context) {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 		return
 	}
-	c.JSON(http.StatusOK, gin.H{"data": rows})
+	c.JSON(http.StatusOK, gin.H{"result": "ok", "data": rows})
 }
 
 func syncTableGetChangesAfter(c *gin.Context) {
@@ -108,7 +108,7 @@ func syncTableGetChangesAfter(c *gin.Context) {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 		return
 	}
-	c.JSON(http.StatusOK, gin.H{"data": rows})
+	c.JSON(http.StatusOK, gin.H{"result": "ok", "data": rows})
 }
 
 func syncTableHandleChanges(c *gin.Context) {
@@ -139,7 +139,7 @@ func syncTableHandleChanges(c *gin.Context) {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 		return
 	}
-	c.JSON(http.StatusOK, gin.H{"data": results})
+	c.JSON(http.StatusOK, gin.H{"result": "ok", "data": results})
 }
 
 func syncTableSyncAt(c *gin.Context) {
@@ -183,7 +183,7 @@ func syncTableSyncAt(c *gin.Context) {
 		ourChange    []sql.SyncRecord
 		done         []bool
 	}{latest, ourChange, done}
-	c.JSON(http.StatusOK, gin.H{"data": ret})
+	c.JSON(http.StatusOK, gin.H{"result": "ok", "data": ret})
 }
 
 func syncTableReplaceAll(c *gin.Context) {
@@ -214,7 +214,7 @@ func syncTableReplaceAll(c *gin.Context) {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 		return
 	}
-	c.JSON(http.StatusOK, gin.H{"data": true})
+	c.JSON(http.StatusOK, gin.H{"result": "ok", "data": true})
 }
 
 func syncTableInsertIfRecent(c *gin.Context) {
@@ -248,7 +248,7 @@ func syncTableInsertIfRecent(c *gin.Context) {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 		return
 	}
-	c.JSON(http.StatusOK, gin.H{"data": done})
+	c.JSON(http.StatusOK, gin.H{"result": "ok", "data": done})
 }
 
 func syncTableInsertOne(c *gin.Context) {
@@ -273,7 +273,7 @@ func syncTableInsertOne(c *gin.Context) {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 		return
 	}
-	c.JSON(http.StatusOK, gin.H{"data": lastModified})
+	c.JSON(http.StatusOK, gin.H{"result": "ok", "data": lastModified})
 }
 
 func syncTableDeleteIfRecent(c *gin.Context) {
@@ -305,7 +305,7 @@ func syncTableDeleteIfRecent(c *gin.Context) {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 		return
 	}
-	c.JSON(http.StatusOK, gin.H{"data": done})
+	c.JSON(http.StatusOK, gin.H{"result": "ok", "data": done})
 }
 
 func syncTableDeleteOne(c *gin.Context) {
@@ -326,5 +326,5 @@ func syncTableDeleteOne(c *gin.Context) {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 		return
 	}
-	c.JSON(http.StatusOK, gin.H{"data": lastModified})
+	c.JSON(http.StatusOK, gin.H{"result": "ok", "data": lastModified})
 }
